+++ src/main/java/vn/edu/vnuk/bnb/rowmapper/BookingsRowMapper.java
@@ -29,13 +29,13 @@ public class BookingsRowMapper implements RowMapper<Booking> {
 		IdentificationType identificationtype = new IdentificationType();
 		RoomType roomtype = new RoomType();
 		
-		country.setId(rs.getLong("country_id"));
+		country.setId(rs.getInt("country_id"));
 		country.setLabel(rs.getString("label"));
 		
-		usertype.setId(rs.getLong("user_type_id"));
+		usertype.setId(rs.getInt("user_type_id"));
 		usertype.setLabel(rs.getString("label"));
 		
-		identificationtype.setId(rs.getLong("identification_type_id"));
+		identificationtype.setId(rs.getInt("identification_type_id"));
 		identificationtype.setLabel(rs.getString("label"));
 		
 		roomtype.setId(rs.getInt("room_type_id"));
@@ -49,8 +49,8 @@ public class BookingsRowMapper implements RowMapper<Booking> {
 		room.setSmoking(rs.getBoolean("is_smoking"));
 		room.setRoomType(roomtype);
 		
-		user.setId(rs.getLong("id"));
-		user.setUserTypesId(rs.getLong("user_type_id"));
+		user.setId(rs.getInt("id"));
+		user.setUserTypesId(rs.getInt("user_type_id"));
 		user.setFirstName(rs.getString("first_name"));
 		user.setMiddleName(rs.getString("middle_name"));
 		user.setLastName(rs.getString("last_name"));
@@ -64,9 +64,9 @@ public class BookingsRowMapper implements RowMapper<Booking> {
 		user.setCountry(country);
 		user.setIdentificationTypes(identificationtype);
 		
-		booking.setId(rs.getLong("id"));
-		booking.setRoomId(rs.getLong("room_id"));
-		booking.setUserId(rs.getLong("user_id"));
+		booking.setId(rs.getInt("id"));
+		booking.setRoomId(rs.getInt("room_id"));
+		booking.setUserId(rs.getInt("user_id"));
 		booking.setCheckIn(rs.getDate("check_in"));
 		booking.setCheckOut(rs.getDate("check_out"));
 		booking.setQuanlity(rs.getInt("quanlity_of_people"));
@@ -96,13 +96,13 @@ public class BookingsRowMapper implements RowMapper<Booking> {
     		roomtype.setId((int) row.get("room_type_id"));
     		roomtype.setLabel((String) row.get("label"));
 			
-    		usertype.setId((Long) row.get("user_type_id"));
+    		usertype.setId((int) row.get("user_type_id"));
     		usertype.setLabel((String) row.get("label"));
 			
-    		identificationtype.setId((Long) row.get("identification_type_id"));
+    		identificationtype.setId((int) row.get("identification_type_id"));
     		identificationtype.setLabel((String) row.get("label"));
 			
-    		country.setId((Long) row.get("country_id"));
+    		country.setId((int) row.get("country_id"));
     		country.setLabel((String) row.get("label"));
 			
 			room.setId((int) row.get("room_id"));
@@ -113,8 +113,8 @@ public class BookingsRowMapper implements RowMapper<Booking> {
 			room.setSmoking((boolean) row.get("is_smoking"));
 			room.setRoomType(roomtype);
 			
-			user.setId((Long) row.get("user_id"));
-			user.setUserTypesId((Long) row.get("user_type_id"));
+			user.setId((int) row.get("user_id"));
+			user.setUserTypesId((int) row.get("user_type_id"));
 			user.setFirstName((String) row.get("first_name"));
 			user.setMiddleName((String) row.get("middle_name"));
 			user.setLastName((String) row.get("last_name"));
@@ -127,9 +127,9 @@ public class BookingsRowMapper implements RowMapper<Booking> {
 			user.setCountry(country);
 			user.setIdentificationTypes(identificationtype);
 			
-			booking.setId((Long) row.get("id"));
-			booking.setRoomId((Long) row.get("room_id"));
-			booking.setUserId((Long) row.get("user_id"));
+			booking.setId((int) row.get("id"));
+			booking.setRoomId((int) row.get("room_id"));
+			booking.setUserId((int) row.get("user_id"));
 			booking.setCheckIn((java.sql.Date) row.get("check_in"));
 			booking.setCheckOut((java.sql.Date) row.get("check_out"));
 			booking.setQuanlity((int) row.get("quanlity_of_people"));
