+++ src/main/java/vn/edu/vnuk/bnb/rowmapper/BookingsServicesRowMapper.java
@@ -33,16 +33,16 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 		Service service = new Service();
 		BookingService bookingservice = new BookingService();
 		
-		country.setId(rs.getLong("country_id"));
+		country.setId(rs.getInt("country_id"));
 		country.setLabel(rs.getString("label"));
 		
-		service.setId(rs.getLong("service_id"));
+		service.setId(rs.getInt("service_id"));
 		service.setLabel(rs.getString("label"));
 		
-		usertype.setId(rs.getLong("user_type_id"));
+		usertype.setId(rs.getInt("user_type_id"));
 		usertype.setLabel(rs.getString("label"));
 		
-		identificationtype.setId(rs.getLong("identification_type_id"));
+		identificationtype.setId(rs.getInt("identification_type_id"));
 		identificationtype.setLabel(rs.getString("label"));
 		
 		roomtype.setId(rs.getInt("room_type_id"));
@@ -56,8 +56,8 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 		room.setSmoking(rs.getBoolean("is_smoking"));
 		room.setRoomType(roomtype);
 		
-		user.setId(rs.getLong("user_id"));
-		user.setUserTypesId(rs.getLong("user_type_id"));
+		user.setId(rs.getInt("user_id"));
+		user.setUserTypesId(rs.getInt("user_type_id"));
 		user.setFirstName(rs.getString("first_name"));
 		user.setMiddleName(rs.getString("middle_name"));
 		user.setLastName(rs.getString("last_name"));
@@ -71,9 +71,9 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 		user.setCountry(country);
 		user.setIdentificationTypes(identificationtype);
 		
-		booking.setId(rs.getLong("booking_id"));
-		booking.setRoomId(rs.getLong("room_id"));
-		booking.setUserId(rs.getLong("user_id"));
+		booking.setId(rs.getInt("booking_id"));
+		booking.setRoomId(rs.getInt("room_id"));
+		booking.setUserId(rs.getInt("user_id"));
 		booking.setCheckIn(rs.getDate("check_in"));
 		booking.setCheckOut(rs.getDate("check_out"));
 		booking.setQuanlity(rs.getInt("quanlity_of_people"));
@@ -81,9 +81,9 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 		booking.setRoom(room);
 		booking.setUser(user);
 		
-		bookingservice.setId(rs.getLong("id"));
-		bookingservice.setBookingId(rs.getLong("booking_id"));
-		bookingservice.setServiceId(rs.getLong("service_id"));
+		bookingservice.setId(rs.getInt("id"));
+		bookingservice.setBookingId(rs.getInt("booking_id"));
+		bookingservice.setServiceId(rs.getInt("service_id"));
 		bookingservice.setPrice(rs.getDouble("price"));
 		bookingservice.setBookings(booking);
 		bookingservice.setServices(service);
@@ -113,17 +113,17 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
     		roomtype.setId((int) row.get("room_type_id"));
     		roomtype.setLabel((String) row.get("label"));
 			
-    		service.setId((Long) row.get("service_id"));
+    		service.setId((int) row.get("service_id"));
     		service.setLabel((String) row.get("label"));
     		
     		
-    		usertype.setId((Long) row.get("user_type_id"));
+    		usertype.setId((int) row.get("user_type_id"));
     		usertype.setLabel((String) row.get("label"));
 			
-    		identificationtype.setId((Long) row.get("identification_type_id"));
+    		identificationtype.setId((int) row.get("identification_type_id"));
     		identificationtype.setLabel((String) row.get("label"));
 			
-    		country.setId((Long) row.get("country_id"));
+    		country.setId((int) row.get("country_id"));
     		country.setLabel((String) row.get("label"));
 			
 			room.setId((int) row.get("room_id"));
@@ -134,8 +134,8 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 			room.setSmoking((boolean) row.get("is_smoking"));
 			room.setRoomType(roomtype);
 			
-			user.setId((Long) row.get("user_id"));
-			user.setUserTypesId((Long) row.get("user_type_id"));
+			user.setId((int) row.get("user_id"));
+			user.setUserTypesId((int) row.get("user_type_id"));
 			user.setFirstName((String) row.get("first_name"));
 			user.setMiddleName((String) row.get("middle_name"));
 			user.setLastName((String) row.get("last_name"));
@@ -148,9 +148,9 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 			user.setCountry(country);
 			user.setIdentificationTypes(identificationtype);
 			
-			booking.setId((Long) row.get("booking_id"));
-			booking.setRoomId((Long) row.get("room_id"));
-			booking.setUserId((Long) row.get("user_id"));
+			booking.setId((int) row.get("booking_id"));
+			booking.setRoomId((int) row.get("room_id"));
+			booking.setUserId((int) row.get("user_id"));
 			booking.setCheckIn((java.sql.Date) row.get("check_in"));
 			booking.setCheckOut((java.sql.Date) row.get("check_out"));
 			booking.setQuanlity((int) row.get("quanlity_of_people"));
@@ -158,9 +158,9 @@ public class BookingsServicesRowMapper implements RowMapper<BookingService> {
 			booking.setRoom(room);
 			booking.setUser(user);
 			
-			bookingservice.setId((Long) row.get("id"));
-			bookingservice.setBookingId((Long) row.get("booking_id"));
-			bookingservice.setServiceId((Long) row.get("service_id"));
+			bookingservice.setId((int) row.get("id"));
+			bookingservice.setBookingId((int) row.get("booking_id"));
+			bookingservice.setServiceId((int) row.get("service_id"));
 			bookingservice.setPrice((double) row.get("price"));
 			booking.setRoom(room);
 			booking.setUser(user);
